+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix `isVerbosityAllowed` default case (log verbosity undefined).
 
 
 ## [0.3.3] - 2020-09-15